requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.7
motor==3.3.2
python-pptx==0.6.23
fpdf==1.7.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0